cleaned up json summary to omit pattern spec if no pattern matching done

+++ shellscripts/networkperturbations_helper_qsub.sh
@@ -72,7 +72,10 @@ fi
 
 # delete intermediate files; it is possible that $STABLEFCLIST does not exist
 rm $NETWORKFILE $DATABASEFILE 
-rm $STABLEFCLIST
+if [ -f $STABLEFCLIST ]; then
+	rm $STABLEFCLIST
+fi
+
